+++ src/ImageList/index.js
@@ -12,7 +12,7 @@ import { RippleFeedback, IconToggle } from '@protonapp/react-native-material-ui'
 import Gradient from './gradient'
 import SearchBarWrapper from '../Shared/SearchWrapper'
 import WrappedIconToggle from '../IconToggle/index.js'
-import EmptyListWrapper from '../Shared/EmptyListWrapper'
+import EmptyState from '../Shared/EmptyState'
 
 export default class ImageList extends Component {
   static defaultProps = {
@@ -143,10 +143,10 @@ export default class ImageList extends Component {
       (items && !items[0]) || openAccordion === 'listEmptyState'
     if (renderEmptyState) {
       return (
-        <EmptyListWrapper
+        <EmptyState
           listEmptyState={listEmptyState}
           items={items}
-        ></EmptyListWrapper>
+        ></EmptyState>
       )
     }
